+++ src/main/java/dabang/star/cafe/infrastructure/service/SHA256EncryptService.java
@@ -1,5 +1,6 @@
-package dabang.star.cafe.domain.login;
+package dabang.star.cafe.infrastructure.service;
 
+import dabang.star.cafe.domain.login.EncryptService;
 import org.springframework.stereotype.Service;
 
 import java.nio.charset.StandardCharsets;
