+++ src/main/java/dabang/star/cafe/infrastructure/service/SessionManagerLoginService.java
@@ -1,6 +1,8 @@
-package dabang.star.cafe.domain.login;
+package dabang.star.cafe.infrastructure.service;
 
 import dabang.star.cafe.api.exception.NoAuthenticationException;
+import dabang.star.cafe.domain.login.EncryptService;
+import dabang.star.cafe.domain.login.ManagerLoginService;
 import dabang.star.cafe.domain.manager.ManagerData;
 import dabang.star.cafe.domain.manager.ManagerRepository;
 import lombok.RequiredArgsConstructor;
