+++ src/main/java/dabang/star/cafe/infrastructure/service/SessionMemberLoginService.java
@@ -1,6 +1,8 @@
-package dabang.star.cafe.domain.login;
+package dabang.star.cafe.infrastructure.service;
 
 import dabang.star.cafe.api.exception.NoAuthenticationException;
+import dabang.star.cafe.domain.login.EncryptService;
+import dabang.star.cafe.domain.login.MemberLoginService;
 import dabang.star.cafe.domain.member.MemberData;
 import dabang.star.cafe.domain.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
